+++ docs/changelog.md
@@ -6,6 +6,21 @@ nav_order: 2
 
 # Changelog
 
+# Unreleased
+
+### Devices
+
+- ExposeSensor: `cooldown` is extended to wait for connection if not established.
+
+### Internals
+
+- Replace `TaskRegistry.register()` with `TaskRegistry.start_task()` for better readability and easier handling.
+- Rename `TaskRegistry.unregister()` to `TaskRegistry.remove_task()`.
+- Allow TaskRegistry Tasks to call regular functions as well as async functions. Rename TaskRegistry.register and Task `async_func` attribute to `target`.
+- Add TaskRegistry Task `wait_before_start` and `wait_for_connection` options to delay task start and wait for an established connection before running the target function.
+- Add TaskRegistry Task `repeat_after` option to automatically restart the task periodically.
+- Remove TaskRegistry.register `track_task` option. All tasks used this option before and it is now the default behaviour.
+
 # 3.14.0 ExposeSensor improvements 2026-01-12
 
 ### Devices